error handling for removing

+++ backend/src/services/removeQuestionAndUserAnswers.js
@@ -6,6 +6,12 @@ export async function removeQuestionAndUserAnswers(questionId) {
   // userAnswers nach der questionId durchsuchen und die gefundenen userAnswers auch löschen?
   // dann die Frage selbst löschen?
 
+  const foundQuestion = await Question.findById(questionId);
+  if (!foundQuestion)
+    throw new Error(
+      "The question with the id " + questionId + " does not exist."
+    );
+
   // userAnswers löschen:
   const removedUserAnswers = await UserAnswer.deleteMany({ questionId });
   // console.log(removedUserAnswers);
